rain is succesfully falling. Working on optimization now.
This commit integrated the fisher yates shuffle and a promise chain in
order to get rain to fall in a randomized order.
I learned a lot about promises and arrays in doing this.
It has taken like two weeks working on this problem and learning all the
ins and outs to make me proficient in this.

 Changes to be committed:
	modified:   css/animate_rain.css
	modified:   js/.main.js.swp
	modified:   js/main.js
	modified:   js/utils.js
	modified:   page1.html